+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix MoveAssetActionProvider to use the new EntitySearch component and resolve a minor warning
 - Fix Sidecar FloatingSidebar to use newly generalized EntityResolver component
 - Fix styling of manage plugins page so 'add plugin' FAB does not obscure plugin tree actions
+- Fix Orbit.js 'some' queries to use Drupal JSON:API 'IN' operator
 
 ## [0.7.1] - 2022-09-09